done with separation
