+++ pages/_document.tsx
@@ -1,3 +1,4 @@
+/** @jsxImportSource @emotion/react */
 import { Html, Head, Main, NextScript } from 'next/document';
 import { css, Global } from '@emotion/react';
 import { global } from './_app.styles';
